Fix RT 107261
This was that code outside any subroutine was being assigned to the
immediately-preceding subroutine, if any.

The 'package'-finding logic was replaced by PPIx::Utilities::Node, and
the all code belonging to that name space was so marked. Then all
subroutines were found, and all code in the subroutine statement and the
subroutine's block was assigned to it.

Options were also added to control the subroutine name returned outside
a subroutine (since main() is a valid subroutine name), and the handling
of fully-qualified subroutine names (sub Foo::Bar {}).

Also tweaked metadata and added author tests.